requirements.txt
mysql-connector-python==9.4.0